Docs pass and organization cleanup
