+++ src/config.rs
@@ -1,5 +1,6 @@
 // HAL configuration (minimal), mirroring embassy-imxrt style
 
+use crate::clocks::config::ClocksConfig;
 use crate::interrupt::Priority;
 
 #[non_exhaustive]
@@ -7,6 +8,7 @@ pub struct Config {
     pub time_interrupt_priority: Priority,
     pub rtc_interrupt_priority: Priority,
     pub adc_interrupt_priority: Priority,
+    pub clock_cfg: ClocksConfig,
 }
 
 impl Default for Config {
@@ -15,6 +17,7 @@ impl Default for Config {
             time_interrupt_priority: Priority::from(0),
             rtc_interrupt_priority: Priority::from(0),
             adc_interrupt_priority: Priority::from(0),
+            clock_cfg: ClocksConfig::default(),
         }
     }
 }
